Create requirement match service (cleanup of CommandExecutorService). Support async context results.
