+++ litecommands-annotations/src/dev/rollczi/litecommands/annotations/MethodDefinition.java
@@ -2,8 +2,8 @@ package dev.rollczi.litecommands.annotations;
 
 import dev.rollczi.litecommands.argument.Argument;
 import dev.rollczi.litecommands.reflect.LiteCommandsReflectInvocationException;
-import dev.rollczi.litecommands.requirement.BindRequirement;
-import dev.rollczi.litecommands.requirement.ContextRequirement;
+import dev.rollczi.litecommands.bind.BindRequirement;
+import dev.rollczi.litecommands.context.ContextRequirement;
 import dev.rollczi.litecommands.requirement.Requirement;
 import dev.rollczi.litecommands.wrapper.WrapFormat;
