+++ litecommands-annotations/src/dev/rollczi/litecommands/annotations/bind/BindRequirementProcessor.java
@@ -1,9 +1,8 @@
 package dev.rollczi.litecommands.annotations.bind;
 
 import dev.rollczi.litecommands.annotations.AnnotationHolder;
-import dev.rollczi.litecommands.annotations.context.Context;
 import dev.rollczi.litecommands.annotations.requirement.RequirementProcessor;
-import dev.rollczi.litecommands.requirement.BindRequirement;
+import dev.rollczi.litecommands.bind.BindRequirement;
 import dev.rollczi.litecommands.requirement.Requirement;
 
 public class BindRequirementProcessor<SENDER> extends RequirementProcessor<SENDER, Bind> {
