+++ litecommands-annotations/src/dev/rollczi/litecommands/annotations/context/ContextRequirementProcessor.java
@@ -2,7 +2,7 @@ package dev.rollczi.litecommands.annotations.context;
 
 import dev.rollczi.litecommands.annotations.AnnotationHolder;
 import dev.rollczi.litecommands.annotations.requirement.RequirementProcessor;
-import dev.rollczi.litecommands.requirement.ContextRequirement;
+import dev.rollczi.litecommands.context.ContextRequirement;
 import dev.rollczi.litecommands.requirement.Requirement;
 
 public class ContextRequirementProcessor<SENDER> extends RequirementProcessor<SENDER, Context> {
