+++ litecommands-core/src/dev/rollczi/litecommands/command/executor/CommandExecutorDefaultReference.java
@@ -5,8 +5,8 @@ import dev.rollczi.litecommands.command.CommandRoute;
 import dev.rollczi.litecommands.meta.Meta;
 import dev.rollczi.litecommands.meta.MetaHolder;
 import dev.rollczi.litecommands.priority.PriorityLevel;
-import dev.rollczi.litecommands.requirement.BindRequirement;
-import dev.rollczi.litecommands.requirement.ContextRequirement;
+import dev.rollczi.litecommands.bind.BindRequirement;
+import dev.rollczi.litecommands.context.ContextRequirement;
 import dev.rollczi.litecommands.requirement.RequirementsResult;
 import java.util.Collections;
 import java.util.List;
