+++ litecommands-core/src/dev/rollczi/litecommands/context/ContextRequirement.java
@@ -1,5 +1,6 @@
-package dev.rollczi.litecommands.requirement;
+package dev.rollczi.litecommands.context;
 
+import dev.rollczi.litecommands.requirement.Requirement;
 import dev.rollczi.litecommands.wrapper.WrapFormat;
 
 import java.util.function.Supplier;
