+++ litecommands-core/src/dev/rollczi/litecommands/context/ContextResult.java
@@ -2,107 +2,51 @@ package dev.rollczi.litecommands.context;
 
 import dev.rollczi.litecommands.requirement.RequirementCondition;
 import dev.rollczi.litecommands.requirement.RequirementFutureResult;
-import dev.rollczi.litecommands.requirement.RequirementResult;
 import java.util.concurrent.CompletableFuture;
+import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.Collections;
 import java.util.List;
 import org.jetbrains.annotations.ApiStatus;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Supplier;
 
-public class ContextResult<T> implements RequirementResult<T>, RequirementFutureResult<T> {
-
-    private final @Nullable Supplier<T> result;
-    private final List<RequirementCondition> conditions;
-    private final Object error;
-
-    private ContextResult(@Nullable Supplier<T> result, Object error, List<RequirementCondition> conditions) {
-        this.result = result;
-        this.error = error;
-        this.conditions = conditions;
-    }
+public interface ContextResult<T> extends RequirementFutureResult<T> {
 
     @Override
-    public @NotNull T getSuccess() {
-        if (result == null) {
-            throw new IllegalStateException("Cannot get success result from failed result");
-        }
+    CompletableFuture<ContextCompletedResult<T>> asFuture();
 
-        return result.get();
+    @ApiStatus.Experimental
+    <R> ContextResult<R> map(Function<T, R> mapper);
+
+    @ApiStatus.Experimental
+    ContextResult<T> mapFailure(Function<Object, ContextResult<T>> mapper);
+
+    @ApiStatus.Experimental
+    <R> ContextResult<R> flatMap(Function<T, ContextResult<R>> mapper);
+
+    @ApiStatus.Experimental
+    ContextResult<T> whenSuccessful(Consumer<T> action);
+
+    @ApiStatus.Experimental
+    ContextResult<T> whenFailed(Consumer<Object> action);
+
+    static <T> ContextCompletedResult<T> ok(Supplier<T> supplier) {
+        return new ContextCompletedResult<>(supplier, null, Collections.emptyList());
     }
 
-    @Override
-    public @NotNull Object getFailedReason() {
-        return error;
-    }
-
-    @Override
-    public @NotNull List<RequirementCondition> getConditions() {
-        return conditions;
-    }
-
-    @Override
-    public boolean isFailed() {
-        return error != null;
-    }
-
-    @Override
-    public boolean isSuccessful() {
-        return result != null;
-    }
-
-    @Override
-    public boolean isSuccessfulNull() {
-        return false;
+    static <T> ContextCompletedResult<T> error(Object error) {
+        return new ContextCompletedResult<>(null, error, Collections.emptyList());
     }
 
     @ApiStatus.Experimental
-    public <R> ContextResult<R> map(Function<T, R> mapper) {
-        if (this.isFailed()) {
-            return ContextResult.error(error);
-        }
-
-        return ContextResult.ok(() -> mapper.apply(getSuccess()));
+    static <T> AsyncContextResult<T> completeableFuture(CompletableFuture<ContextResult<T>> future) {
+        return new AsyncContextResult<>(future);
     }
 
     @ApiStatus.Experimental
-    public <R> ContextResult<R> flatMap(Function<T, ContextResult<R>> mapper) {
-        if (this.isFailed()) {
-            return ContextResult.error(error);
-        }
-
-        return mapper.apply(getSuccess());
-    }
-
-    public static <T> ContextResult<T> ok(Supplier<T> supplier) {
-        return new ContextResult<>(supplier, null, Collections.emptyList());
-    }
-
-    public static <T> ContextResult<T> error(Object error) {
-        return new ContextResult<>(null, error, Collections.emptyList());
-    }
-
-    @ApiStatus.Experimental
-    public static <T> ContextResult<T> conditional(Supplier<T> supplier, List<RequirementCondition> conditions) {
-        return new ContextResult<>(supplier, null, Collections.unmodifiableList(conditions));
-    }
-
-    @Override
-    public CompletableFuture<RequirementResult<T>> asFuture() {
-        return CompletableFuture.completedFuture(this);
-    }
-
-    @Override
-    public RequirementResult<T> asResultOr(RequirementResult<T> result) {
-        return this;
-    }
-
-    @Override
-    public RequirementResult<T> await() {
-        return this;
+    static <T> ContextResult<T> conditional(Supplier<T> supplier, List<RequirementCondition> conditions) {
+        return new ContextCompletedResult<>(supplier, null, Collections.unmodifiableList(conditions));
     }
 
 }
