+++ litecommands-core/src/dev/rollczi/litecommands/requirement/RequirementFutureResult.java
@@ -6,10 +6,36 @@ import org.jetbrains.annotations.ApiStatus;
 @ApiStatus.Experimental
 public interface RequirementFutureResult<PARSED> {
 
-    CompletableFuture<RequirementResult<PARSED>> asFuture();
+    CompletableFuture<? extends RequirementResult<PARSED>> asFuture();
 
-    RequirementResult<PARSED> asResultOr(RequirementResult<PARSED> result);
+    @SuppressWarnings("unchecked")
+    default RequirementResult<PARSED> getNow(RequirementResult<PARSED> valueIfAbsent) {
+        if (this instanceof RequirementResult) {
+            return (RequirementResult<PARSED>) this;
+        }
 
-    RequirementResult<PARSED> await();
+        CompletableFuture<? extends RequirementResult<PARSED>> completableFuture = asFuture();
+
+        if (completableFuture.isCancelled() || completableFuture.isCompletedExceptionally()) {
+            return valueIfAbsent;
+        }
+
+        RequirementResult<PARSED> now = completableFuture.getNow(null);
+
+        if (now == null) {
+            return valueIfAbsent;
+        }
+
+        return now;
+    }
+
+    @SuppressWarnings("unchecked")
+    default RequirementResult<PARSED> await()  {
+        if (this instanceof RequirementResult) {
+            return (RequirementResult<PARSED>) this;
+        }
+
+        return asFuture().join();
+    }
 
 }
