+++ litecommands-programmatic/src/dev/rollczi/litecommands/programmatic/LiteCommand.java
@@ -10,8 +10,8 @@ import dev.rollczi.litecommands.join.JoinArgument;
 import dev.rollczi.litecommands.meta.Meta;
 import dev.rollczi.litecommands.meta.MetaKey;
 import dev.rollczi.litecommands.quoted.QuotedStringArgumentResolver;
-import dev.rollczi.litecommands.requirement.BindRequirement;
-import dev.rollczi.litecommands.requirement.ContextRequirement;
+import dev.rollczi.litecommands.bind.BindRequirement;
+import dev.rollczi.litecommands.context.ContextRequirement;
 import dev.rollczi.litecommands.scheduler.SchedulerPoll;
 import dev.rollczi.litecommands.strict.StrictMode;
 import dev.rollczi.litecommands.wrapper.WrapFormat;
